Split code and prevent creation of multiple quotes
